requirements.txt
fastapi
uvicorn
undetected-chromedriver
selenium
python-dotenv
cachetools